Parallelized the DP write process.
Previously, every generation automatically runs all of the cases and all of the populations.
This is fixed by only running the ones that are in the current generation.
This only matters for the first 2 runs. First run GA will only run one population and the second run GA will run the rest of the first generation.
Still this should save considerable time.

Main issue is the adjustment of the range of fitting parameters.
Most of the long-hour runs were due to runs non-converging.
Most problematic issues are the gamma_b and A_fact terms for backstress.
Also adjusted gDD for martensite from 1e-3 -> 40e-3 to 10e-3 -> 40e-3 to avoid issues with going across an order of magnitude.